requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==2.0.12
Django==3.2.7
djangorestframework==3.12.4
idna==3.4
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0
requests==2.26.0
sqlparse==0.4.4
urllib3==1.26.16
